refactor: extract formatYearMonth/formatAmount to server.util
