+++ server/src/main/kotlin/server/money/MoneyWebhookService.kt
@@ -22,6 +22,7 @@ import server.util.DISCORD_EMBED_COLOR
 import server.util.WebhookServiceType
 import server.util.await
 import server.util.detectWebhookService
+import server.util.formatYearMonth
 
 class MoneyWebhookService(
     private val firestore: Firestore,
@@ -129,15 +130,6 @@ class MoneyWebhookService(
 
     companion object {
         private val appUrl: String? = EnvConfig["APP_URL"]
-
-        /** "YYYY-MM" を "YYYY年MM月" 表記（月は 0 埋め 2 桁）に整形。パース失敗時は入力をそのまま返す。 */
-        internal fun formatYearMonth(yearMonth: String): String {
-            val parts = yearMonth.split("-")
-            if (parts.size != 2) return yearMonth
-            val year = parts[0].toIntOrNull() ?: return yearMonth
-            val month = parts[1].toIntOrNull() ?: return yearMonth
-            return "%d年%02d月".format(year, month)
-        }
     }
 }
