+++ server/src/test/kotlin/server/money/MoneyWebhookPayloadTest.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
-import server.money.MoneyWebhookService.Companion.formatYearMonth
+import server.util.formatYearMonth
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertIs
